backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3
redis==5.1.1
aiohttp==3.10.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
boto3==1.35.30
pandas==2.2.3
numpy==2.1.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
